+++ tool_animation/map_animator.py
@@ -39,8 +39,8 @@ def get_layer_by_id(layer_id):
 def strip_agg_options(param: str) -> str:
     for opt in AGGREGATION_OPTIONS:
         if param.endswith("_" + opt):
-            return param.rstrip("_" + opt)
-
+            stripchars = 1 + len(opt)
+            return param[:-stripchars]
     return param
 
 
@@ -337,26 +337,27 @@ class MapAnimator(QGroupBox):
         self._update_parameter_attributes()
 
         Q_CUM = 'q_cum'
-        default_items = {WATERLEVEL.name, Q_CUM}
+        default_short_names = {WATERLEVEL.name, Q_CUM}
         if Q_CUM not in (v['parameters'] for v in self.line_parameters.values()):
-            default_items.add(DISCHARGE.name)
+            default_short_names.add(DISCHARGE.name)
 
         for combo_box, parameters in (
             (self.line_parameter_combo_box, self.line_parameters),
             (self.node_parameter_combo_box, self.node_parameters),
         ):
-            last_param = combo_box.currentText()
+            current_long_name = combo_box.currentText()
             combo_box.clear()
             if parameters:
-                items = sorted(parameters)
-                combo_box.addItems(items)
-                if last_param in items:
+                long_names = sorted(parameters)
+                combo_box.addItems(long_names)
+                try:
                     # keep last choice if possible
-                    new_index = items.index(last_param)
-                else:
+                    new_index = long_names.index(current_long_name)
+                except ValueError:
                     # activate a default item if possible
-                    for index, item in enumerate(items):
-                        if item in default_items:
+                    for index, long_name in enumerate(long_names):
+                        short_name = parameters[long_name]["parameters"]
+                        if short_name in default_short_names:
                             new_index = index
                             break
                     else:  # no break, activate the first item
@@ -452,9 +453,13 @@ class MapAnimator(QGroupBox):
             timestep_nr = bisect_left(parameter_timestamps, current_seconds)
             timestep_nr = min(timestep_nr, parameter_timestamps.size - 1)
 
-            values_t0 = threedi_result.get_values_by_timestep_nr(parameter, 0)
+            # get the data
+            ids = self._get_feature_ids(layer)
+            values_t0 = threedi_result.get_values_by_timestep_nr(
+                variable=parameter, timestamp_idx=0, node_ids=ids,
+            )
             values_ti = threedi_result.get_values_by_timestep_nr(
-                parameter, timestep_nr
+                variable=parameter, timestamp_idx=timestep_nr, node_ids=ids,
             )
 
             # theedigrid may have returned masked arrays in the past
@@ -468,6 +473,7 @@ class MapAnimator(QGroupBox):
                 values_t0[values_t0 == NO_DATA_VALUE] = np.NaN
                 values_ti[values_ti == NO_DATA_VALUE] = np.NaN
 
+            # determine which fields to update
             ti_field_index, t0_field_index = (
                 layer.fields().indexOf(n)
                 for n in result_item._result_field_names[layer_id]
@@ -475,30 +481,15 @@ class MapAnimator(QGroupBox):
             assert ti_field_index != -1
             assert t0_field_index != -1
 
-            ids = self._get_feature_ids(layer)
-
-            # NOTE OF CAUTION: subtracting 1 from id  is mandatory for
-            # groundwater because those indexes start from 1 (something to
-            # do with a trash element), but for the non-groundwater version
-            # it is not. HOWEVER, due to some magic hackery in how the
-            # *_result layers are created/copied from the regular result
-            # layers, the resulting feature ids also start from 1, which
-            # why we need to subtract it in both cases, which btw is
-            # purely coincidental.
-            # TODO: to avoid all this BS this part should be refactored
-            # by passing the index to get_values_by_timestep_nr, which
-            # should take this into account
-            # print(min(ids), max(ids), values_t0.size, values_ti.size)
-            dvalues_t0 = values_t0[ids - 1]
-            dvalues_ti = values_ti[ids - 1]
+            # update layer
             update_dict = {
                 k: {
                     t0_field_index: NULL if math.isnan(v0) else v0,
                     ti_field_index: NULL if math.isnan(vi) else vi,
                 } for k, v0, vi in zip(
                     ids.tolist(),
-                    dvalues_t0.tolist(),
-                    dvalues_ti.tolist(),
+                    values_t0.tolist(),
+                    values_ti.tolist(),
                 )
             }
             provider.changeAttributeValues(update_dict)